refactor: code cleanup, consolidate non-streaming handlers and make errorHandler cross-VM compatible
