+++ .antigravitycli/6e487c10-0ebe-408a-a95f-46c56632517e.json
@@ -1,0 +1,1 @@
+/Users/c/.gemini/config/projects/6e487c10-0ebe-408a-a95f-46c56632517e.json
