+++ src/routes/responses.ts
@@ -16,7 +16,9 @@ import { summarizeRequestForLog } from "../logs/request-summary.js";
 import { getRealClientIp } from "../utils/get-real-client-ip.js";
 import { randomUUID } from "crypto";
 import { getConfig } from "../config.js";
-import { prepareSchema } from "../translation/shared-utils.js";
+import { extractProxyApiKey } from "../utils/extract-api-key.js";
+import { errorHandler } from "../middleware/error-handler.js";
+import { prepareSchema, isRecord } from "../translation/shared-utils.js";
 import { parseModelName, resolveModelId, buildDisplayModelName } from "../models/model-store.js";
 import { handleProxyRequest } from "./shared/proxy-handler.js";
 import { handleDirectRequest } from "./shared/direct-request-handler.js";
@@ -27,7 +29,7 @@ import {
   OPENAI_SUBAGENT_HEADER,
   sanitizeClientMetadata,
 } from "../proxy/openai-subagent.js";
-import { isRecord, PASSTHROUGH_FORMAT } from "./responses-passthrough.js";
+import { PASSTHROUGH_FORMAT } from "./responses-passthrough.js";
 import { handleCompact } from "./responses-compact.js";
 
 // Re-export for downstream consumers
@@ -75,8 +77,7 @@ function checkAuth(
 
   const config = getConfig();
   if (config.server.proxy_api_key) {
-    const authHeader = c.req.header("Authorization");
-    const providedKey = authHeader?.replace(/^bearer\s+/i, "");
+    const providedKey = extractProxyApiKey(c);
     if (!providedKey || !accountPool.validateProxyApiKey(providedKey)) {
       c.status(401);
       return c.json({
@@ -116,22 +117,10 @@ export function createResponsesRoutes(
   upstreamRouter?: UpstreamRouter,
 ): Hono {
   const app = new Hono();
+  app.onError(errorHandler);
 
   const responsesHandler = async (c: Context) => {
-    let rawBody: unknown;
-    try {
-      rawBody = await c.req.json();
-    } catch {
-      c.status(400);
-      return c.json({
-        type: "error",
-        error: {
-          type: "invalid_request_error",
-          code: "invalid_json",
-          message: "Malformed JSON request body",
-        },
-      });
-    }
+    const rawBody = await c.req.json();
 
     const body = parseBody(c, rawBody);
     if (body instanceof Response) return body;
@@ -297,20 +286,7 @@ export function createResponsesRoutes(
   };
 
   const compactHandler = async (c: Context) => {
-    let rawBody: unknown;
-    try {
-      rawBody = await c.req.json();
-    } catch {
-      c.status(400);
-      return c.json({
-        type: "error",
-        error: {
-          type: "invalid_request_error",
-          code: "invalid_json",
-          message: "Malformed JSON request body",
-        },
-      });
-    }
+    const rawBody = await c.req.json();
 
     const body = parseBody(c, rawBody);
     if (body instanceof Response) return body;
