+++ tests/e2e/chat.test.ts
@@ -64,7 +64,7 @@ function buildApp(opts?: { noAccount?: boolean }): TestContext {
 
   const app = new Hono();
   app.use("*", requestId);
-  app.use("*", errorHandler);
+  app.onError(errorHandler);
   app.route("/", createChatRoutes(accountPool, cookieJar, proxyPool));
   app.route("/", createModelRoutes());
   app.route("/", createWebRoutes(accountPool));
