+++ tests/e2e/gemini.test.ts
@@ -57,7 +57,7 @@ function buildApp(opts?: { noAccount?: boolean }): TestContext {
 
   const app = new Hono();
   app.use("*", requestId);
-  app.use("*", errorHandler);
+  app.onError(errorHandler);
   app.route("/", createGeminiRoutes(accountPool, cookieJar, proxyPool));
   app.route("/", createModelRoutes());
   app.route("/", createWebRoutes(accountPool));
