+++ tests/e2e/quota-refresh.test.ts
@@ -47,7 +47,7 @@ beforeAll(() => {
 
   app = new Hono();
   app.use("*", requestId);
-  app.use("*", errorHandler);
+  app.onError(errorHandler);
   app.route("/", createAccountRoutes(pool, scheduler));
 });
