+++ tests/e2e/responses.test.ts
@@ -63,7 +63,7 @@ function buildApp(opts?: { noAccount?: boolean }): TestContext {
 
   const app = new Hono();
   app.use("*", requestId);
-  app.use("*", errorHandler);
+  app.onError(errorHandler);
   app.route("/", createResponsesRoutes(accountPool, cookieJar, proxyPool));
   app.route("/", createModelRoutes());
   app.route("/", createWebRoutes(accountPool));
