+++ tests/unit/routes/shared/non-streaming-empty-response-exhausted.test.ts
@@ -1,5 +1,5 @@
 import type { AccountPool } from "@src/auth/account-pool.js";
-import { handleNonStreamingEmptyResponseExhausted } from "@src/routes/shared/non-streaming-empty-response-exhausted.js";
+import { handleNonStreamingEmptyResponseExhausted } from "@src/routes/shared/non-streaming-helpers.js";
 import type { ProxyRequest } from "@src/routes/shared/proxy-handler-types.js";
 import { describe, expect, it, vi } from "vitest";
